# api/docs/swagger.yaml
definitions:
  auth_service.CreateUser:
    properties:
      email:
        type: string
      key:
        type: string
      name:
        type: string
      secret:
        type: string
    type: object
  auth_service.LoginRequest:
    properties:
      name:
        type: string
      secret:
        type: string
    type: object
  auth_service.TokenResponse:
    properties:
      Token:
        type: string
    type: object
  auth_service.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      key:
        type: string
      name:
        type: string
      secret:
        type: string
      updated_at:
        type: string
    type: object
  auth_service.UserListResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/auth_service.User'
        type: array
    type: object
  book_service.Book:
    properties:
      author:
        type: string
      cover:
        type: string
      created_at:
        type: string
      id:
        type: string
      isbn:
        type: string
      pages:
        type: integer
      published:
        type: string
      status:
        description: 0-new, 1-reading, 2-finished,
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  book_service.BookListResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/book_service.Book'
        type: array
      count:
        type: integer
    type: object
  book_service.CreateBook:
    properties:
      author:
        type: string
      cover:
        type: string
      isbn:
        type: string
      pages:
        type: integer
      published:
        type: string
      status:
        description: 0-new, 1-reading, 2-finished,
        type: integer
      title:
        type: string
    type: object
  book_service.UpdateBook:
    properties:
      author:
        type: string
      cover:
        type: string
      id:
        type: string
      isbn:
        type: string
      pages:
        type: integer
      published:
        type: string
      status:
        description: 0-new, 1-reading, 2-finished,
        type: integer
      title:
        type: string
    type: object
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
  models.UpdatePatchRequest:
    properties:
      fields:
        additionalProperties: true
        type: object
      id:
        type: string
    type: object
info:
  contact: {}
paths:
  /books:
    get:
      consumes:
      - application/json
      description: Get Book List
      operationId: get_book_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: BookListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/book_service.BookListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Book List
      tags:
      - Book
    post:
      consumes:
      - application/json
      description: Create Book
      operationId: create_book
      parameters:
      - description: CreateBookRequestBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/book_service.CreateBook'
      produces:
      - application/json
      responses:
        "200":
          description: GetBookBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/book_service.Book'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Book
      tags:
      - Book
  /books/:title:
    get:
      consumes:
      - application/json
      description: Get Book By title
      operationId: get_book_by_title
      parameters:
      - description: title
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: BookBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/book_service.Book'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Book By title
      tags:
      - Book
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Book
      operationId: delete_book
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Book
      tags:
      - Book
    patch:
      consumes:
      - application/json
      description: Patch Book
      operationId: patch_book
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdatePatchRequestBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Book data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/book_service.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Patch Book
      tags:
      - Book
    put:
      consumes:
      - application/json
      description: Update Book
      operationId: update_book
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateBookRequestBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/book_service.UpdateBook'
      produces:
      - application/json
      responses:
        "200":
          description: Book data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/book_service.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Book
      tags:
      - Book
  /login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: LoginRequestBody
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth_service.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login
      tags:
      - Authentication
  /myself:
    get:
      consumes:
      - application/json
      description: Get Myself
      operationId: get_user_myself
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/auth_service.User'
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Myself
      tags:
      - User
  /signup:
    post:
      consumes:
      - application/json
      description: Signup
      operationId: Signup
      parameters:
      - description: CreateUserRequest
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Signup
      tags:
      - Authentication
  /user:
    get:
      consumes:
      - application/json
      description: Get User List
      operationId: get_user_list
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserListResponse
          schema:
            $ref: '#/definitions/auth_service.UserListResponse'
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get User List
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create User
      operationId: create_user
      parameters:
      - description: CreateUserRequestBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: UserResponse
          schema:
            $ref: '#/definitions/auth_service.User'
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      summary: Create User
      tags:
      - User
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Get User By ID
      operationId: get_user_by_id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserResponse
          schema:
            $ref: '#/definitions/auth_service.User'
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      summary: Get User By ID
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"